fix: polish project icon editor
