+++ AlasTests/DialogChromeLayoutTests.swift
@@ -1,0 +1,9 @@
+import Testing
+@testable import Alas
+
+struct DialogChromeLayoutTests {
+    @Test("project dialogs reserve enough width for icon controls")
+    func projectDialogWidthFitsIconControls() {
+        #expect(DialogContainerLayout.projectWidth >= 640)
+    }
+}
