+++ AlasTests/SpaceEmojiPickerSelectionTests.swift
@@ -2,12 +2,12 @@ import Testing
 @testable import Alas
 
 @MainActor
-struct SpaceEmojiPickerSelectionTests {
+struct EmojiPickerSelectionTests {
     @Test func validEmojiSelectionCommitsAndDismissesPicker() {
         var selections: [String] = []
         var dismissCount = 0
 
-        let didCommit = SpaceEmojiPickerSelection.commit(
+        let didCommit = EmojiPickerSelection.commit(
             "🚀",
             onSelect: { selections.append($0) },
             dismissPicker: { dismissCount += 1 }
@@ -22,7 +22,7 @@ struct SpaceEmojiPickerSelectionTests {
         var selections: [String] = []
         var dismissCount = 0
 
-        let didCommit = SpaceEmojiPickerSelection.commit(
+        let didCommit = EmojiPickerSelection.commit(
             "work",
             onSelect: { selections.append($0) },
             dismissPicker: { dismissCount += 1 }
